Add delegate for FixedPagingViewController
Add delegate that is called when transitioning to a ViewControllerItem
with its index in the menu items. Allows you to track which view
controller that is currently or about to be selected.
