+++ Parchment/Classes/FixedPagingViewController.swift
@@ -1,23 +1,14 @@
 import UIKit
 
-public struct ViewControllerItem: PagingTitleItem, Equatable {
-  
-  public let viewController: UIViewController
-  public let title: String
-  
-  init(viewController: UIViewController) {
-    self.viewController = viewController
-    self.title = viewController.title ?? ""
-  }
-}
-
-public func ==(lhs: ViewControllerItem, rhs: ViewControllerItem) -> Bool {
-  return lhs.viewController == rhs.viewController
+public protocol FixedPagingViewControllerDelegate : class {
+  func fixedPagingViewController(fixedPagingViewController: FixedPagingViewController, willScrollToItem: ViewControllerItem, atIndex index: Int)
+  func fixedPagingViewController(fixedPagingViewController: FixedPagingViewController, didScrollToItem: ViewControllerItem, atIndex index: Int)
 }
 
 open class FixedPagingViewController: PagingViewController<ViewControllerItem> {
   
-  let items: [ViewControllerItem]
+  open let items: [ViewControllerItem]
+  open weak var itemDelegate: FixedPagingViewControllerDelegate?
   
   public init(viewControllers: [UIViewController], options: PagingOptions = DefaultPagingOptions()) {
     items = viewControllers.map { ViewControllerItem(viewController: $0) }
@@ -31,6 +22,28 @@ open class FixedPagingViewController: PagingViewController<ViewControllerItem> {
 
   required public init?(coder: NSCoder) {
       fatalError("init(coder:) has not been implemented")
+  }
+
+  // MARK: EMPageViewControllerDelegate
+  
+  open override func em_pageViewController(_ pageViewController: EMPageViewController, didFinishScrollingFrom startingViewController: UIViewController?, destinationViewController: UIViewController, transitionSuccessful: Bool) {
+    super.em_pageViewController(pageViewController, didFinishScrollingFrom: startingViewController, destinationViewController: destinationViewController, transitionSuccessful: transitionSuccessful)
+    
+    if let index = items.index(where: { $0.viewController == destinationViewController }) {
+      itemDelegate?.fixedPagingViewController(
+        fixedPagingViewController: self,
+        didScrollToItem: items[index],
+        atIndex: index)
+    }
+  }
+  
+  public func em_pageViewController(_ pageViewController: EMPageViewController, willStartScrollingFrom startingViewController: UIViewController, destinationViewController: UIViewController) {
+    if let index = items.index(where: { $0.viewController == destinationViewController }) {
+      itemDelegate?.fixedPagingViewController(
+        fixedPagingViewController: self,
+        willScrollToItem: items[index],
+        atIndex: index)
+    }
   }
   
 }
